# swagger.json
{
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This package provides a REST API to the Ignition Fuel server.",
    "title": "Ignition Fuel Server REST API",
    "contact": {
      "email": "info@openrobotics.org"
    },
    "license": {
      "name": "Apache 2.0"
    },
    "version": "0.1.0"
  },
  "host": "fuel.ignitionrobotics.org",
  "basePath": "/1.0",
  "paths": {
...
    "/{username}/likes/models": {
      "get": {
        "description": "Get a list of models liked by the specified user.\nModels will be returned paginated, with pages of 20 models by default.\nThe user can request a different page with query parameter 'page' (first page is value 1).\nThe page size can be controlled with query parameter 'per_page', with a maximum of\n100 items per page.\nThe route supports the 'order' parameter, with values 'asc' and 'desc' (default: desc).\nIt also supports the 'q' parameter to perform a fulltext search on models name,\ndescription and tags.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "tags": [
          "models"
        ],
        "summary": "Get models liked by a user.",
        "operationId": "modelLikeList",
        "responses": {
          "200": {
            "$ref": "#/responses/jsonModels"
          },
          "default": {
            "$ref": "#/responses/fuelError"
          }
        }
      }
    },
...
    "/{username}/models/{model}/{version}/files": {
      "get": {
        "description": "Return the files information of a given model.",
        "produces": [
          "application/json",
          "application/x-protobuf"
        ],
        "schemes": [
          "https"
        ],
        "tags": [
          "models"
        ],
        "summary": "Model's file tree.",
        "operationId": "modelFileTree",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "UUID",
            "name": "uuid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ModelFileTree"
          },
          "default": {
            "$ref": "#/responses/fuelError"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ErrMsg": {
      "description": "ErrMsg is serialized as JSON, and returned if the request does not succeed\nTODO: consider making ErrMsg an 'error'",
      "type": "object",
      "properties": {
        "errcode": {
          "description": "Internal error code.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ErrCode"
        },
        "errid": {
          "description": "Generated ID for easy tracking in server logs",
          "type": "string",
          "x-go-name": "ErrID"
        },
        "extra": {
          "description": "Extra information/arguments associated to Error message.",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Extra"
        },
        "msg": {
          "description": "Error message.",
          "type": "string",
          "x-go-name": "Msg"
        },
        "remote-address": {
          "description": "Associated request remote address, if applicable",
          "type": "string",
          "x-go-name": "RemoteAddress"
        },
        "route": {
          "description": "Associated request Route, if applicable",
          "type": "string",
          "x-go-name": "Route"
        },
        "user-agent": {
          "description": "Associated request User-Agent, if applicable",
          "type": "string",
          "x-go-name": "UserAgent"
        }
      },
      "x-go-package": "gitlab.com/ignitionrobotics/web/ign-go"
    },
...
    "Model": {
      "description": "Model base model definition, including fields `ID`, `CreatedAt`, `UpdatedAt`, `DeletedAt`, which could be embedded in your models\ntype User struct {\ngorm.Model\n}",
      "type": "object",
      "properties": {
        "CreatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "DeletedAt": {
          "type": "string",
          "format": "date-time"
        },
        "ID": {
          "type": "integer",
          "format": "uint64"
        },
        "UpdatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "x-go-package": "github.com/jinzhu/gorm"
    },
...
  },
  "responses": {
...
    "fuelError": {
      "description": "Ign Fuel error serialized as JSON",
      "schema": {
        "$ref": "#/definitions/ErrMsg"
      }
    },
    "jsonModels": {
      "description": "Array of Models",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/Model"
        }
      }
    }
  }
}